188 ux improvements home page (#225)
* Modified a few icons to be less misleading

* Icon for hiding side-menu now dynamical

Icon/Button for hiding the side-menu was (imo) very misleading. Changed the triple bars to dynamical arrow. If supposed to close, arrow points left, if to open, arrow points right.

* Fix the Height of the Box and make more than 5 trainings visible

table scrollable

* Time added

Time and Translation for Time (Zeit) added

* Time added - expanded to english translation

* Reset Height

* Highlight newest Dataset

RowStyleFunc - newest Dataset has now orange background

* Availabe Dataset-Types now with Filetype-Info

This change required modifcations (or rather additions) to the Ontology-File. Changes still have to be given for approval to Alexander.

* Made the highlight a little less flashy

* Removing the ?=? idea

As the introduction of subcomments in the comment was declined, removing the changes. Kept the method for splitting up the comment (maybe it could be reused for something), but feel free to delete it completely.

Co-authored-by: AnnaHlzhsr <anna.holzheuser@stud.h-da.de>