Stabilized WikiDataGrabber
Added a few failsafes so that the program won't immediately crash if the entries are not perfect.
Added a few documentation comments.
Renamed returned variable in GetDataFrom(), as not always wikipediaContent is returned.
Added a few TODOs, @AnnaHlzhsr please solve them, you'll probably know better than me.
Currently NO IMAGES are shown, as I hardcoded a pointless string (otherwise the Program crashes, part of the TODOs)